+++ src/interfaces/IChildERC20.sol
@@ -9,7 +9,7 @@ interface IChildERC20 {
 
     function bridge() external view returns (address);
 
-    function parent() external view returns (address token, uint16 chainId);
+    function parent() external view returns (address token, uint64 chainId);
 
     function mint(address to, uint amount) external;
